+++ hardware/ford_hybrid_handler.py
@@ -11,7 +11,7 @@ import time
 from collections import deque
 
 logger = logging.getLogger('openTPT.ford_hybrid')
-from hardware.hardware_handler_base import BoundedQueueHardwareHandler
+from utils.hardware_base import BoundedQueueHardwareHandler
 from utils.config import FORD_HYBRID_ENABLED, FORD_HYBRID_CHANNEL, FORD_HYBRID_BITRATE